Fix problem with facebook request params.
Facebook adds a request param 'fbclid=xxxx' to link in facebook messenger.

The current Django app does not handle this correctly (for Safari on Mac at least).
This leads to a 404.

This could be fixed in Django, but that's more work to find out exactly where to fix it.
Furthermore, this needs a rebuild of the Django image.

The Cleo app uses nginx. It turns out that nginx can check for these fbclid query params.
If they occur, a redirect can be done to the url, without the fbclid param.